requirements.txt
pydub~=0.25.1
openai~=1.61.1
python-dotenv~=1.0.1
elevenlabs~=1.52.0